Fix some pylint complaints.
* Remove unnecessary imports.
* Move shared fixtures to conftest.py.
* Remove unnecessary lambdas.
